+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+* Aurora Postgres construct
+* allow customization of alb health_check_path
+* ElastiCache Cluster construct
+
 # 3.1.0
 
 * ALB: Requiring CIDR range for AlbIngressCidr parameter